- refactored study
