+++ song-python-sdk/tests/tests.py
@@ -26,6 +26,8 @@ from overture_song.entities import *
 from overture_song.model import *
 from overture_song.tools import FileUploadState
 from overture_song.utils import *
+from overture_song.validation import DataField, validation
+from dataclasses import *
 
 
 class TestFile(object):
